# openapi-example-shoppingcart/server/src/main/java/com/ionos/demo/shoppingcart/openapi/api/CartApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (5.2.0).
 * https://openapi-generator.tech Do not edit the class manually.
 */
package com.ionos.demo.shoppingcart.openapi.api;

import com.ionos.demo.shoppingcart.openapi.model.ShoppingCartItem;
import com.ionos.demo.shoppingcart.openapi.model.ShoppingCartObject;
import io.swagger.annotations.*;
import java.util.List;
import java.util.Optional;
import javax.validation.constraints.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
@Validated
@Api(value = "cart", description = "the cart API")
public interface CartApi {

  default Optional<NativeWebRequest> getRequest() {
    return Optional.empty();
  }

  /**
   * DELETE /cart/{cartId} Deletes the shopping cart.
   *
   * @param cartId (required)
   * @return The shopping cart service response. (status code 204)
   */
  @ApiOperation(
      value = "",
      nickname = "deleteCartById",
      notes = "Deletes the shopping cart. ",
      response = ShoppingCartObject.class,
      tags = {
        "User",
        "Middleware",
      })
  @ApiResponses(
      value = {
        @ApiResponse(
            code = 204,
            message = "The shopping cart service response.",
            response = ShoppingCartObject.class)
      })
  @DeleteMapping(
      value = "/cart/{cartId}",
      produces = {"application/json"})
  default ResponseEntity<ShoppingCartObject> deleteCartById(
      @ApiParam(value = "", required = true) @PathVariable("cartId") String cartId) {
    getRequest()
        .ifPresent(
            request -> {
              for (MediaType mediaType : MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                  String exampleString =
                      "{ \"owner\" : \"stephan.fuhrmann\", \"expires_at\" :"
                          + " \"2000-01-23T04:56:07.000+00:00\", \"created_at\" :"
                          + " \"2000-01-23T04:56:07.000+00:00\", \"id\" : \"id\" }";
                  ApiUtil.setExampleResponse(request, "application/json", exampleString);
                  break;
                }
              }
            });
    return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
  }

  /**
   * GET /cart/{cartId} Gets the shopping cart.
   *
   * @param cartId (required)
   * @return The shopping cart service response. (status code 200)
   */
  @ApiOperation(
      value = "",
      nickname = "getCartById",
      notes = "Gets the shopping cart. ",
      response = ShoppingCartObject.class,
      tags = {
        "User",
        "Middleware",
      })
  @ApiResponses(
      value = {
        @ApiResponse(
            code = 200,
            message = "The shopping cart service response.",
            response = ShoppingCartObject.class)
      })
  @GetMapping(
      value = "/cart/{cartId}",
      produces = {"application/json"})
  default ResponseEntity<ShoppingCartObject> getCartById(
      @ApiParam(value = "", required = true) @PathVariable("cartId") String cartId) {
    getRequest()
        .ifPresent(
            request -> {
              for (MediaType mediaType : MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                  String exampleString =
                      "{ \"owner\" : \"stephan.fuhrmann\", \"expires_at\" :"
                          + " \"2000-01-23T04:56:07.000+00:00\", \"created_at\" :"
                          + " \"2000-01-23T04:56:07.000+00:00\", \"id\" : \"id\" }";
                  ApiUtil.setExampleResponse(request, "application/json", exampleString);
                  break;
                }
              }
            });
    return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
  }

  /**
   * GET /cart/{cartId}/items Gets the shopping cart items.
   *
   * @param cartId (required)
   * @return The shopping cart service response. (status code 200)
   */
  @ApiOperation(
      value = "",
      nickname = "getCartItems",
      notes = "Gets the shopping cart items. ",
      response = ShoppingCartItem.class,
      responseContainer = "List",
      tags = {
        "User",
        "Middleware",
      })
  @ApiResponses(
      value = {
        @ApiResponse(
            code = 200,
            message = "The shopping cart service response.",
            response = ShoppingCartItem.class,
            responseContainer = "List")
      })
  @GetMapping(
      value = "/cart/{cartId}/items",
      produces = {"application/json"})
  default ResponseEntity<List<ShoppingCartItem>> getCartItems(
      @ApiParam(value = "", required = true) @PathVariable("cartId") String cartId) {
    getRequest()
        .ifPresent(
            request -> {
              for (MediaType mediaType : MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                  String exampleString =
                      "{ \"quantity\" : 1, \"index\" : 0, \"item_asin\" : \"B0711Z67W9\" }";
                  ApiUtil.setExampleResponse(request, "application/json", exampleString);
                  break;
                }
              }
            });
    return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
  }

  /**
   * PUT /cart/{cartId}/items Updates the shopping cart items.
   *
   * @param cartId (required)
   * @return The shopping cart service response. (status code 200)
   */
  @ApiOperation(
      value = "",
      nickname = "putCartItems",
      notes = "Updates the shopping cart items. ",
      response = ShoppingCartItem.class,
      responseContainer = "List",
      tags = {
        "User",
        "Middleware",
      })
  @ApiResponses(
      value = {
        @ApiResponse(
            code = 200,
            message = "The shopping cart service response.",
            response = ShoppingCartItem.class,
            responseContainer = "List")
      })
  @PutMapping(
      value = "/cart/{cartId}/items",
      produces = {"application/json"})
  default ResponseEntity<List<ShoppingCartItem>> putCartItems(
      @ApiParam(value = "", required = true) @PathVariable("cartId") String cartId) {
    getRequest()
        .ifPresent(
            request -> {
              for (MediaType mediaType : MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                  String exampleString =
                      "{ \"quantity\" : 1, \"index\" : 0, \"item_asin\" : \"B0711Z67W9\" }";
                  ApiUtil.setExampleResponse(request, "application/json", exampleString);
                  break;
                }
              }
            });
    return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
  }
}